Reorganize a bit
